- aplica ajustes na entrada de validacao dos endpoints.

+++ app.py
@@ -11,24 +11,28 @@ from decimal import *
 app= Flask(__name__)
 
 def valida_header(autenticacao):
+
     if not autenticacao:
         resposta = {
             "mensagem": "Erro de autenticação",
             "detalhe" : "Não foi encontrada informações de autenticação"
         }
-        return make_response('Acesso não autorizado', 401)
+        return {'validado': False,'resposta': make_response(resposta, 401)}
     
     if not autenticacao == ("4b9f2a87-704e-4685-b0bf-668b1c5eb8b1"):
         resposta = {
             "mensagem": "Erro de autenticação",
             "detalhe": "Erro ao validar o token"
         }
-        return make_response('Token de autenticação não encontrado na base de dados', 401)
+        return {'validado': False,'resposta': make_response(resposta, 401)}
     
 # Request para adicionar filme
 @app.post("/filme")
 def Add_filme():
-    valida_header(request.headers.get('Authorization'))
+    
+    validacao = valida_header(request.headers.get('Authorization'))
+    if not (validacao['validado']):
+        return validacao['resposta']
     
     filme_enviado = request.get_json()
     id_do_filme = FilmeService.add_filme(filme_enviado['titulo'],filme_enviado['diretor'],filme_enviado['ano_lancamento'],filme_enviado['genero'],filme_enviado['nota'])
@@ -43,7 +47,10 @@ def Add_filme():
 # Pegar somente um filme
 @app.get("/filmes/<int:id>")
 def get_filme_by_id(id):
-    valida_header(request.headers.get('Authorization'))
+    
+    validacao = valida_header(request.headers.get('Authorization'))
+    if not (validacao['validado']):
+        return validacao['resposta']
 
     try:
         filme = FilmeService.get_filme_by_id(id)
@@ -56,14 +63,22 @@ def get_filme_by_id(id):
 # Criando endpoint para selecionar todos os filmes salvos
 @app.get("/filmes")
 def get_all_films():
-    valida_header(request.headers.get('Authorization'))
+    
+    validacao = valida_header(request.headers.get('Authorization'))
+    if not (validacao['validado']):
+        return validacao['resposta']
+
     dados_filmes = FilmeService.get_filmes()
     return make_response(jsonify(dados_filmes),200)
 
 # Modificar as informações pegando o id
 @app.put("/filmes/<id>")
 def atualizar_filme(id):
-    valida_header(request.headers.get('Authorization'))
+    
+    validacao = valida_header(request.headers.get('Authorization'))
+    if not (validacao['validado']):
+        return validacao['resposta']
+
     try:
         data = request.get_json()
         filme_atualizado = FilmeService.update_filme(
@@ -84,13 +99,21 @@ def atualizar_filme(id):
 # Deletar filme
 @app.delete("/filmes/<int:id>")
 def deletar_filme(id):
-    valida_header(request.headers.get('Authorization'))
+    
+    validacao = valida_header(request.headers.get('Authorization'))
+    if not (validacao['validado']):
+        return validacao['resposta']
+
     FilmeService.delete_filme(id)
     return ("Filme deletado com sucesso")
 
 @app.post("/nota")
 def Add_nota():
-    valida_header(request.headers.get('Authorization'))
+    
+    validacao = valida_header(request.headers.get('Authorization'))
+    if not (validacao['validado']):
+        return validacao['resposta']
+
 
     nota_enviada= request.get_json()
     id_do_filme = Notas.add_notas(nota_enviada['id_filme'],nota_enviada['nota'])
@@ -102,13 +125,21 @@ def Add_nota():
 
 @app.get("/notas/<int:idFilme>") #retornar
 def get_filmes(idFilme):
-    valida_header(request.headers.get('Authorization'))
+    
+    validacao = valida_header(request.headers.get('Authorization'))
+    if not (validacao['validado']):
+        return validacao['resposta']
+
     notas = Notas.get_nota_by_id_filme(idFilme)
     return make_response(notas,200)
 
 @app.get("/filme/nota/<int:idFilme>")
 def get_avaliacao_filme_by_id(idFilme):
-    valida_header(request.headers.get('Authorization'))
+    
+    validacao = valida_header(request.headers.get('Authorization'))
+    if not (validacao['validado']):
+        return validacao['resposta']
+
     # pegar todas as notas do filme
     notas = Notas.get_nota_by_id_filme(idFilme)
     info_filme =  FilmeService.get_filme_by_id(idFilme)
